Move loaders out of finetune, which is only for training, while loader used for generation too
